pyproject.toml
[tool.poetry]
name = "metis-job"
version = "0.1.0"
description = ""
authors = ["Col Perks <wild.fauve@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
metis-fn = "^0.1.0"
pino = "^0.6.0"
pendulum = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-env = "^1.1.3"
pyspark = "^3.5.1"
delta-spark = "^3.1.0"
bpython = "^0.24"
pdbpp = "^0.10.3"
bevy = "^2.0.2"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
python_classes = [
    "*Test",
    "Test*",
    "Describe*",
]
python_functions = [
    "test_*",
    "it_*",
]
xfail_strict = true
log_cli = true
log_cli_level = 20
testpaths = [
    "tests",
    "integration",
]

[tool.pytest_env]
ENVIRONMENT = "test"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
